Major updates to the Config handler and installer
I rewrote the Get, Set and DB init functions of the ConfigHandler; it now uses ConfigModels instead of manually creating datamodels as needed.  This allows for more information to be tracked internally and for easier Configuration creation, (ie: during installation).

Fixed installation, it now works again because of the aforementioned config updates.

Updated all usages of ConfigHandler::GetValue to use the new slightly-shortened version ConfigHandler::Get.