Updates for the the Page system
Including adding a grouped where clause into the dataset system.  This allows WHERE statements to contain sub groups joined by custom operators, (AND, OR, etc).
Added instantiated method as a verification method.  This is useful if the model needs to call an internal method for validation due to data constraints or any other reason.  To use this, set the validation to array('this', 'nameofmethod').
Model::validate() will also return the string of the validation error if there is one, (unless $throwexception is requested).
Signed-off-by:cdp1337 <powellc@powelltechs.com>