More rearranging, yayz!
Moved the theme bits into its own component.  This simplifies the code a little and makes it more atomic.
Started to set the system license to AGPL, though references have not been updated completely yet.
Moved some of the utilities into their own directory.  These are not crucial to the application and will be distributed in a separate development release.