More tweaks...
Continuing development on the Theme system, specifically widgets.
Updated the Core Navigation system, now only one page is saved per
controller/method.  This enables a controller to simply redirect to the
baseurl of the requested controller and all core and GET parameters are
appended along with it.
Tweaks to the View class, added request and response arrays, to contain
their respective sets of data.