Bug #127
Models now translate the values being sent in with the "set" function. This is useful with Booleans, where setting "false" may result in a 1 in the database, since a non-typecasted comparison would evaluate it as a non-zero length string.

Currently, this only affects boolean keys that have not had their "set" function overwritten.