Updates to the core
Setting package as 2.1.4 version
Added a "/core/page/postrender" hook.  This fires after all the other page logic has executed.
Redirect component errors to the error log instead of to the screen.
Fix in the View for if the controller method returns an invalid return.
Removed the "last-modified" meta tag in favor of the article og version.  last-modified is no longer valid as of HTML5.
Fix in the mysqli backend for managing the schemas.  There was a bug if a primary key did
not exist, then was created later.
Fix in the pagemodel where rewriteurl initialized value would be blank on new models.
Fix on 404 pages, now the system's view will not be lost.