Performance tuning.
There is now a compiler, which takes the bootstrap file and its immediate dependencies and "compiles" it into a single file.  This saves PHP from having to open each file individually upon loading the site, and saves approximately 20 - 40 ms on a SSD, presumably more on a traditional HDD that has a slower seek time.