Core work
Overhauled the template system entirelly.  Template is now a factory-only class and the specific rendering engines are now polled automatically based on the file extension.
Changed the Core::GoBack method to ignore the same page that the user is currently on.  This allows "1" to be used on a POST page as if it were the original page load.
bug Make sure not to record the navigation for 404 pages.
feature New support for optional CSS files in template files.