Fix file inputs with no value selected by default, but not required.
There was a minor bug with file inputs that had no value, but weren't required.
The upload option was being checked by default in this case, which caused issues to the user if ignored,
because the submitting logic saw that a file upload was requested, but no value set and threw an error.