Change the terminology for creating a new blog listing page.
Now that Core displays all create-page links in a central location, it was confusing if people read "Create Blog Page",
when it should have read "Create Blog Listing Page".

You can have a blog page without the blog listing structure, now that a blog is simply a type of content.