Better support for URL overriding in CLI mode.
The site URL can now be set and will be respected when in CLI mode
via the core settings 'site_url' or the $_SERVER['HTTP_HOST']
property when that's not set.

This doesn't really do anything by itself, but allows getURL()
to return expected results when generated from CLI scripts.